created first draft of GUI
main GUI and 3 other files, two of which are secondary windows to be used to fulfil requirements of the task, the third is to define styles for ease of application across the 3 GUI files.

Currently, the GUI lacks functionality, but this will be remedied hopefully tomorrow. Of course, this requires knowledge of what AI models are being used, so let me know what's going on and what outputs you need for whatever AI model