Updated GUI
Made various QOL changes to secondary_window.py that also just happen to align with the assignment's OOP requirements. such as:
- i have separated the Transient status to it's own class which i can then apply to secondary windows as i see fit
- i have included 2 decorator functions to validate that the secondary window has a parent and to automatically focus on the new window

Additionally i have added content to the InfoWindow - it now displays information on OOP techniques used as of yet.

Finally, i have added functionality to the HELP button on the menubar, it now shows a prompt with a temporary help message that will be changed in later versions